+++ source/libruntime/private/callstack.posix.cpp
@@ -15,7 +15,7 @@ auto up::callstack::readTrace(span<uintptr> addresses, uint skip) -> span<uintpt
     void* buffer = nullptr;
 
     uint max = addresses.size() - min<uint>(addresses.size(), skip);
-    uint count = backtrace(&buffer, max);
+    uint count = backtrace(&buffer, static_cast<int>(max));
     skip = min(skip, count);
 
     std::memcpy(addresses.data(), static_cast<uintptr*>(buffer) + skip, min(count, max));
@@ -28,7 +28,7 @@ auto up::callstack::resolveTraceRecords(span<uintptr const> addresses, span<Trac
     uint max = addresses.size() < records.size() ? addresses.size() : records.size();
 
     void* const addrs = const_cast<uintptr*>(addresses.data()); // NOLINT(cppcoreguidelines-pro-type-const-cast)
-    char** symbols = backtrace_symbols(&addrs, addresses.size());
+    char** symbols = backtrace_symbols(&addrs, static_cast<int>(addresses.size()));
 
     for (auto index : sequence(max)) {
         auto& record = records[index];
