+++ source/spud/public/potato/spud/_hash_table.h
@@ -18,9 +18,9 @@
 
 namespace up::_detail::hash_table {
     struct constants {
-        static constexpr uint8 SPECIAL = 0xff;
-        static constexpr uint8 EMPTY = 0x80;
-        static constexpr uint8 TOMBSTONE = 0xfe;
+        static constexpr int8 SPECIAL = static_cast<int8>(0xff);
+        static constexpr int8 EMPTY = static_cast<int8>(0x80);
+        static constexpr int8 TOMBSTONE = static_cast<int8>(0xfe);
 
         static constexpr size_t SENTINEL = ~size_t{0};
 
@@ -42,10 +42,10 @@ namespace up::_detail::hash_table {
     };
 
     struct match_ops_sse {
-        static inline unsigned match(uint8 const* control, uint8 match) noexcept;
-        static inline unsigned matchEmpty(uint8 const* control) noexcept;
-        static inline unsigned matchEmptyOrTombstone(uint8 const* control) noexcept;
-        static inline unsigned matchFull(uint8 const* control) noexcept;
+        static inline unsigned match(int8 const* control, int8 match) noexcept;
+        static inline unsigned matchEmpty(int8 const* control) noexcept;
+        static inline unsigned matchEmptyOrTombstone(int8 const* control) noexcept;
+        static inline unsigned matchFull(int8 const* control) noexcept;
     };
 
     template <typename Item>
@@ -55,86 +55,84 @@ namespace up::_detail::hash_table {
         static constexpr size_t padding(size_t groups) noexcept;
         static constexpr size_t size(size_t groups) noexcept;
 
-        static void allocate(size_t groups, uint8*& out_control, Item*& out_items);
+        static void allocate(size_t groups, int8*& out_control, Item*& out_items);
         static void deallocate(size_t groups, void* data);
     };
 
     template <typename Hash>
     struct hash_ops {
         static constexpr Hash _h1(Hash hash) noexcept { return hash >> constants::H2_SHIFT; }
-        static constexpr uint8 _h2(Hash hash) noexcept { return static_cast<uint8>(hash & constants::H1_MASK); }
+        static constexpr int8 _h2(Hash hash) noexcept { return static_cast<int8>(hash & constants::H1_MASK); }
     };
 
     template <typename Item, typename Key, typename Hash, typename Equality, typename HashOps, typename MatchOps>
     struct table_ops {
         static constexpr size_t find(
             size_t groups,
-            uint8 const* control,
+            int8 const* control,
             Item const* items,
             Key const& key,
             hash_result_t<Hash, Key> hash) noexcept;
         static constexpr size_t findEmptyOrTombstone(
             size_t groups,
-            uint8 const* control,
+            int8 const* control,
             hash_result_t<Hash, Key> hash) noexcept;
         static constexpr bool erase(
             size_t groups,
-            uint8* control,
+            int8* control,
             Item* items,
             Key const& key,
             hash_result_t<Hash, Key> hash) noexcept;
 
         // returns true if the caller must drop allocated memory (it won't be cleared/reset)
-        static bool clear(size_t groups, uint8* control, Item* items) noexcept;
+        static bool clear(size_t groups, int8* control, Item* items) noexcept;
     };
 
-    unsigned match_ops_sse::match(uint8 const* control, uint8 match) noexcept {
+    unsigned match_ops_sse::match(int8 const* control, int8 match) noexcept {
         auto const* const slots = reinterpret_cast<__m128i const*>(control);
         return _mm_movemask_epi8(_mm_cmpeq_epi8(_mm_loadu_si128(slots), _mm_set1_epi8(match)));
     }
 
-    unsigned match_ops_sse::matchEmpty(uint8 const* control) noexcept {
+    unsigned match_ops_sse::matchEmpty(int8 const* control) noexcept {
         auto const* const slots = reinterpret_cast<__m128i const*>(control);
         auto const mask = _mm_loadu_si128(slots);
         return _mm_movemask_epi8(_mm_sign_epi8(mask, mask));
     }
 
-    unsigned match_ops_sse::matchEmptyOrTombstone(uint8 const* control) noexcept {
+    unsigned match_ops_sse::matchEmptyOrTombstone(int8 const* control) noexcept {
         auto const* const slots = reinterpret_cast<__m128i const*>(control);
         return _mm_movemask_epi8(_mm_cmpgt_epi8(_mm_set1_epi8(constants::SPECIAL), _mm_loadu_si128(slots)));
     }
 
-    unsigned match_ops_sse::matchFull(uint8 const* control) noexcept {
-        return ~matchEmptyOrTombstone(control) & 0xffff;
-    }
+    unsigned match_ops_sse::matchFull(int8 const* control) noexcept { return ~matchEmptyOrTombstone(control) & 0xffff; }
 
     template <typename Item>
     constexpr auto memory_ops<Item>::padding(size_t groups) noexcept -> size_t {
-        size_t const control = sizeof(uint8) * groups * constants::WIDTH;
+        size_t const control = sizeof(int8) * groups * constants::WIDTH;
         return control % alignof(Item);
     }
 
     template <typename Item>
     constexpr auto memory_ops<Item>::size(size_t groups) noexcept -> size_t {
-        size_t const control = sizeof(uint8) * groups * constants::WIDTH;
+        size_t const control = sizeof(int8) * groups * constants::WIDTH;
         size_t const items = sizeof(Item) * groups * constants::WIDTH;
         return control + items + padding(groups);
     }
 
     template <typename Item>
-    void memory_ops<Item>::allocate(size_t groups, uint8*& out_control, Item*& out_items) {
+    void memory_ops<Item>::allocate(size_t groups, int8*& out_control, Item*& out_items) {
         using constants = constants;
 
         auto const bytes = size(groups);
         auto const slots = groups * constants::WIDTH;
-        auto const itemsOffset = sizeof(uint8) * slots + padding(groups);
+        auto const itemsOffset = sizeof(int8) * slots + padding(groups);
 
         void* const data = operator new(bytes, std::align_val_t(alignment));
 
-        out_control = static_cast<uint8*>(data);
+        out_control = static_cast<int8*>(data);
         out_items = reinterpret_cast<Item*>(static_cast<char*>(data) + itemsOffset);
 
-        std::memset(out_control, constants::EMPTY, sizeof(uint8) * slots);
+        std::memset(out_control, constants::EMPTY, sizeof(int8) * slots);
     }
 
     template <typename Item>
@@ -145,7 +143,7 @@ namespace up::_detail::hash_table {
     template <typename Item, typename Key, typename Hash, typename Equality, typename HashOps, typename MatchOps>
     constexpr auto table_ops<Item, Key, Hash, Equality, HashOps, MatchOps>::find(
         size_t groups,
-        uint8 const* control,
+        int8 const* control,
         Item const* items,
         Key const& key,
         hash_result_t<Hash, Key> hash) noexcept -> size_t {
@@ -184,7 +182,7 @@ namespace up::_detail::hash_table {
     template <typename Item, typename Key, typename Hash, typename Equality, typename HashOps, typename MatchOps>
     constexpr auto table_ops<Item, Key, Hash, Equality, HashOps, MatchOps>::findEmptyOrTombstone(
         size_t groups,
-        uint8 const* controls,
+        int8 const* controls,
         hash_result_t<Hash, Key> hash) noexcept -> size_t {
         size_t group = HashOps::_h1(hash) & (groups - 1);
         size_t probe = 1;
@@ -207,7 +205,7 @@ namespace up::_detail::hash_table {
     template <typename Item, typename Key, typename Hash, typename Equality, typename HashOps, typename MatchOps>
     constexpr bool table_ops<Item, Key, Hash, Equality, HashOps, MatchOps>::erase(
         size_t groups,
-        uint8* control,
+        int8* control,
         Item* items,
         Key const& key,
         hash_result_t<Hash, Key> hash) noexcept {
@@ -237,7 +235,7 @@ namespace up::_detail::hash_table {
     template <typename Item, typename Key, typename Hash, typename Equality, typename HashOps, typename MatchOps>
     bool table_ops<Item, Key, Hash, Equality, HashOps, MatchOps>::clear(
         size_t groups,
-        uint8* control,
+        int8* control,
         Item* items) noexcept {
         using constants = constants;
 
@@ -277,7 +275,7 @@ namespace up::_detail::hash_table {
         // only clear control if our size is small, otherwise clearing
         // is more expensive than just re-allocating
         if (!shouldDrop) {
-            std::memset(control, constants::EMPTY, sizeof(uint8) * groups * constants::WIDTH);
+            std::memset(control, constants::EMPTY, sizeof(int8) * groups * constants::WIDTH);
         }
 
         return shouldDrop;
